+++ bgstally/state.py
@@ -25,7 +25,8 @@ class State:
         self.ShowZeroActivitySystems:tk.StringVar = tk.StringVar(value=config.get_str('XShowZeroActivity', default=CheckStates.STATE_ON))
         self.AbbreviateFactionNames:tk.StringVar = tk.StringVar(value=config.get_str('XAbbreviate', default=CheckStates.STATE_OFF))
         self.IncludeSecondaryInf:tk.StringVar = tk.StringVar(value=config.get_str('XSecondaryInf', default=CheckStates.STATE_ON))
-        self.DiscordWebhook:tk.StringVar = tk.StringVar(value=config.get_str('XDiscordWebhook', default=""))
+        self.DiscordBGSWebhook:tk.StringVar = tk.StringVar(value=config.get_str('XDiscordWebhook', default=""))
+        self.DiscordFCJumpWebhook:tk.StringVar = tk.StringVar(value=config.get_str("XDiscordFCJumpWebhook", default=""))
         self.DiscordUsername:tk.StringVar = tk.StringVar(value=config.get_str('XDiscordUsername', default=""))
         self.EnableOverlay:tk.StringVar = tk.StringVar(value=config.get_str('XEnableOverlay', default=CheckStates.STATE_ON))
 
@@ -58,7 +59,8 @@ class State:
         config.set('XShowZeroActivity', self.ShowZeroActivitySystems.get())
         config.set('XAbbreviate', self.AbbreviateFactionNames.get())
         config.set('XSecondaryInf', self.IncludeSecondaryInf.get())
-        config.set('XDiscordWebhook', self.DiscordWebhook.get())
+        config.set('XDiscordWebhook', self.DiscordBGSWebhook.get())
+        config.set('XDiscordFCJumpWebhook', self.DiscordFCJumpWebhook.get())
         config.set('XDiscordUsername', self.DiscordUsername.get())
         config.set('XEnableOverlay', self.EnableOverlay.get())